requirements.txt
asgiref==3.3.1
Django==3.1.7
django-admin-rangefilter==0.8.1
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
xlwt==1.3.0
